[emasamplerate] Fix inaccuracies with low throughputs/small counts
Adapting the AvgSampleRate implementation, which uses integers, to the Exponential Moving Average, which uses floats, introduced some inaccuracies with very small counts, which end up in the EMA at < 1. This results in their log being < 0, and skews `logSum` to be smaller than it would have been computed in the AvgSampleRate implementation. All this adds up to less accurate sample rate calculations when counts are small or there are a lot of keys with counts < 1. This change puts a floor on the EMA counts at 1 for the purposes of doing the log curve mapping.

I added two tests:
- TestEMASampleUpdateMapsSparseCounts
- TestAvgSampleUpdateMapsSparseCounts

These should be equal, but before my EMASampleRate changes, the EMA test got a sample rate of 5 while the AvgSampleRate test got a sample rate value of 16.

Note that no changes occurred in the expected output of `TestEMASampleUpdateMaps`, which further suggests this was an edge case only affecting small counts.