feat: EMAThroughput sampler (#58)
This adds a new sampler type, an EMAThroughput sampler.

The EMASampleRate sampler requires specifying a desired sample rate. But
more often, what someone would like to do is specify a given throughput.
The WindowedThroughput sampler does that, but does so at the cost of
maintaining a buffer of N previous sample rates, each of which is of
equal weight.

The EMAThroughput sampler maintains a diminishing moving average of
samples over time; older data is of less weight than current data. It
also doesn't need to maintain any extra buffers.

Fixes #51.