feat: Add metrics counter retrieval (#65)
## Which problem is this PR solving?

To date, the samplers have had no way of getting any metrics from them,
which has sometimes made it hard to support. This adds a simple
GetMetrics call that retrieves a map of named metrics on request.

GetMetrics returns a map of metrics about the sampler's performance. All
values are returned as int64; counters are cumulative and the names
always end with "_count", while gauges are instantaneous with no
particular naming convention. All names are prefixed with the string
passed in.

## Short description of the changes

- Add GetMetrics function to Sampler interface and all samplers